Added stub wrapper code
This is the first commit of the wrapper library. With the exception
of root.rs, all code was [generated from the JSON structs using an
ad-hoc program](https://gist.github.com/alteous/bdc4bcc9ce8949838f01c30aae323c64)
and hand-tweaked in order to compile. The stub code is deliberately
incomplete to allow for multiple candidate implementations.