Fixed invalid mesh export when 'primitives' is empty.
The GLTF standard requires every mesh to contain a `primitives` property. Also,
several importers, such as Blender's GLTF import, fail when a mesh does not contain
the `primitives` property.

By skipping the `primitives` property when it is empty, exporting empty meshes will
result in an invalid GLTF file. This change removes this attribute so that the `primitives`
property is always exported, even when it is empty.

Honestly, this such a minor detail that it's probably not even worth worrying
about, as it won't even affect the vast majority of people. The only reason I
even noticed this is because I'm writing a modding tool for a game that, for some
reason, has empty meshes in certain levels.