Add test for size hint on sparse accessor
Currently this fails due to a faulty size_hint() implementation in
SparseIter.

For the test asset it claims that 3 items will be read from the
iterator, when the correct answer is 14. 3 is the number of positions
overriden by the sparse values, while the total number of positions is
14.

The bug itself will be fixed in an upcoming change.