Fix bug in size hint of sparse iterator
The fix is to delegate size hint to `self.base` as it knows how many
values are left to consume.

Note that `base` is only set if `accessor.bufferView` is set. If it
isn't then the sparse iterator actually has no clue on how many items
are left. But it at least knows that there are at least `self.values`
items left.
This issue will be improved in a follow up PR.

The problem with the old code was that when `self.values.len()`
is less than `self.counter`, a panic was triggered:
`attempt to subtract with overflow`.

To understand why we need to understand what the `values` and `counter`
variables.

`self.values` is an iterator over the sparse values that overwrites
values from the base accessor. Each call to `SparseIter::next()` will
consume one item and `self.values.len()` consequently decrease by 1.

`self.counter` holds how many items have been consumed or seen another
way the number of successfull calls to `SparseIter::next()`. It starts
at 0 and increase until all values in the base accessor has been
consumed.

Now you hopefully see that the old implementation was just plain wrong.