Enable reading sparse accessor without buffer view
When `accessor.bufferView` is unset the sparse iterator should
use T::zero() for `accessor.count` items as base. Then possibly
overwriting them with the values in specified by the accessor.sparse
section.

The sparse iterator is passed `accessor.count` here, so that it knows
when it should stop generating items. Before it continued until
the end of times.